% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHitFaRates.R
\name{calcHitFaRates}
\alias{calcHitFaRates}
\title{Calculating hit and false alarm rates}
\usage{
calcHitFaRates(
  oldRating,
  ratings,
  oldNew,
  oldNewLevels = c(0, 1),
  transformation = TRUE
)
}
\arguments{
\item{oldRating}{Values corresponding to 'old' responses.}

\item{ratings}{A vector containing all responses.}

\item{oldNew}{An vector containing whether an item was new/not-studied or old/studied.}

\item{oldNewLevels}{An vector containing  possible levels of old_newInformation. The first value or level is for new/not-studied 
items and the second value is for old/studied items. Defaults to c(0,1).}

\item{transformation}{Logical indicating to choose a log-linear transformation following Snodgrass & Corwin (1988). Default is TRUE.}
}
\value{
The function returns a data frame with hitRate and falseAlarm_rate.
}
\description{
This function calculates hit and false alarm rates to be used by dPrime and responseBias. The default it to apply log-linear transformed percentage values.
}
\references{
Snodgrass, J. G., & Corwin, J. (1988). Pragmatics of measuring recognition memory: Applications to dementia and amnesia. Journal of Experimental Psychology: General, 117(1), 34–50. https://doi.org/10.1037/0096-3445.117.1.34
}
\author{
Joern Alexander Quent, \email{alexander.quent@rub.de}
}
\keyword{ROC}
\keyword{alarm}
\keyword{false}
\keyword{hit}
\keyword{memory}
