% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DPSD.R
\name{DPSD}
\alias{DPSD}
\title{Estimation of recollection and familiarity by fitting to Dual Process Signal Detection (DPSD) model}
\usage{
DPSD(responseScale, confidenceRatings, oldNew, oldNewLevels = c(0, 1))
}
\arguments{
\item{responseScale}{An vector containing  possible levels of recognition responses ordered from highest to lowest (e.g. 6:1).}

\item{confidenceRatings}{An vector containing recognition responses according to levels of the variable responseScale.}

\item{oldNew}{An vector coding whether an item was new/not-studied or old/studied.}

\item{oldNewLevels}{An vector containing  possible levels of old_newInformation. The first value or level is for new/not-studied 
items and the second value is for old/studied items. Defaults to c(0,1).}
}
\value{
The function the set of parameters, which showed the lowest total squared difference:
\item{recollection}{Estimate of recollection.}  
\item{familiarity}{Estimate of familiarity.}
}
\description{
This function allows to estimate recollection and familiarity by fitting data to the DPSD model. 
The optimization is attempted by minimizing the total squared difference between observed and predicted hit and false 
alarm rates. The Broyden-Fletcher-Goldfarb-Shanno (BFGS) algorithm from the function.
}
\author{
Joern Alexander Quent, \email{alexander.quent@rub.de}
}
\references{
Yonelinas, A. P. (1994). Receiver-operating characteristics in recognition memory: evidence for a dual-process model. Journal of Experimental Psychology: Learning, Memory, and Cognition, 20(6), 1341.
}
\keyword{DPSD}
\keyword{ROC}
\keyword{familiarity}
\keyword{recollection}

