% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitDPSD.R
\name{fitDPSD}
\alias{fitDPSD}
\title{Estimation of recollection and familiarity by fitting to Dual Process Signal Detection (DPSD) model}
\usage{
fitDPSD(falseAlarm, hit, iterations = 200, startValues = c(0.2, 0.5,
  seq(0.5, -1.5, length = length(falseAlarm))))
}
\arguments{
\item{falseAlarm}{A vector containing the false alarm rate.}

\item{hit}{A vector containing the hit rate.}

\item{iterations}{A numeric value specifying the number of iterations. Default is set to 200, which can take a little bit of time.}

\item{startValues}{bla}
}
\value{
The function returns a list with components:
\item{recollection}{The median of the estimations recollection.}  
\item{familiarity}{The median of the estimations of familiarity.}
}
\description{
This function allows to estimate recollection and familiarity by fitting it to DPSD model. 
The optimization is attempted by minimizing the total squared difference between observed and predicted hit and false 
alarm rates. The Broyden-Fletcher-Goldfarb-Shanno (BFGS) algorithm from the function. 
\code{\link{optim}} is used for optimization. Box constraints limit estimations of recollection and familirity to be positive.
}
\author{
Joern Alexander Quent, \email{alexander.quent@rub.de}
}
\references{
Yonelinas, A. P. (1994). Receiver-operating characteristics in recognition memory: evidence for a dual-process model. Journal of Experimental Psychology: Learning, Memory, and Cognition, 20(6), 1341.
}
\seealso{
A matlab implementation by Dorian Pustina served as an example for this function. 
See \code{\link{github.com/dorianps/memorysolve}}.
}
\keyword{ROC}
\keyword{familiarity}
\keyword{recollection}

