% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rememberKnow.R
\name{rememberKnow}
\alias{rememberKnow}
\title{Estimation of recollection and familiarity using the ROC Remember/Know procedure}
\usage{
rememberKnow(rememberLevels, knowLevels, confidenceRatings, oldNew,
  oldNewLevels = c(0, 1))
}
\arguments{
\item{rememberLevels}{A value indicating the code for the remember response.}

\item{knowLevels}{A vector contatining the codes for the know responses.}

\item{confidenceRatings}{An vector containing recognition responses coding remember and know responses.}

\item{oldNew}{An vector coding whether an item was new/not-studied or old/studied.}

\item{oldNewLevels}{An vector containing  possible levels of old_newInformation. The first value or level is for new/not-studied 
items and the second value is for old/studied items. Defaults to c(0,1).}
}
\value{
The function returns a list with components:
\item{recollection}{The estimation of recollection.}
\item{familiarity}{The estimation of familiarity.}
}
\description{
This function allows to estimate recollection and familiarity using Remember/Know procedure. 
In this procedure, recollection is given by the probability that an old/studied item were 
given a remember response, while familiarity is given by the probility that an old/studied 
item were given a know response divived by the probility that an old/studied item were not 
given a remember response. 
\deqn{recollection = P(remember)} 
\deqn{familiarity = P(know)/(1 -P(remember))}
}
\details{
The value of \code{\link{recollection}} is represented by the probability of a remember response for old items. The value of 
\code{\link{familiarity}} is represented by the probability for a know response for old items divided by 1  - \code{\link{recollection}}.
}
\author{
Joern Alexander Quent, \email{alexander.quent@rub.de}
}
\references{
Yonelinas, A. P. (2001). Consciousness, Control, and Confidence: The 3 Cs of Recognition Memory. Journal of Experimental Psychology. General, 130(3), 361.
}
\keyword{ROC}
\keyword{Remember/Know}
\keyword{familiarity}
\keyword{memory}
\keyword{recollection}

