% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/returnFittedROC.R
\name{returnFittedROC}
\alias{returnFittedROC}
\title{Returning false alarm and hit rates for the Dual Process Signal Detection (DPSD)}
\usage{
returnFittedROC(recollection, familiarity)
}
\arguments{
\item{recollection}{A value representing the recollection rate.}

\item{familiarity}{A value representing the familiarity rate}
}
\value{
The function returns a list with components:
\item{falseAlarm}{The fitted false alarm rate.}
\item{hit}{The fitted hit rate.}
}
\description{
This function allows you to get the correspending false alarm and hit rates for a given set of
recollection and familiarity assuming that the variance of the old item distirbution is 1.
}
\author{
Joern Alexander Quent, \email{alexander.quent@rub.de}
}
\references{
Yonelinas, A. P., & Parks, C. M. (2007). Receiver operating characteristics (ROCs) in recognition memory: A review. Psychological Bulletin, 133(5), 800 - 832. http://doi.org/10.1037/0033-2909.133.5.800
}
\keyword{ROC}
\keyword{memory}

