% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rocAUC.R
\name{rocAUC}
\alias{rocAUC}
\title{Caculation of Area Under the Curve (AUC) for an ROC plot}
\usage{
rocAUC(falseAlarm, hit)
}
\arguments{
\item{falseAlarm}{A vector containing the false alarm rate.}

\item{hit}{A vector containing the hit rate.}
}
\value{
The function returns the estimation of AUC.
}
\description{
This function calculates the AUC by summing the tri- and rectangles, which can be made of the ROC points. 
If missing, the y-intercept is added by linear interpolation. The last point, where both false Alarm rates reach 1, is also added.
}
\author{
Joern Alexander Quent, \email{alexander.quent@rub.de}
}
\keyword{AUC}
\keyword{Area}
\keyword{Curve}
\keyword{ROC}
\keyword{Under}
\keyword{memory}

